Update README file and refactor various aspects of codebase
In this commit, the readme.md file is significantly updated, providing an extensive usage and development guide. Additionally, extracted the 'scripts' constant to its own file, 'scripts.ts', in order to improve modularity. Some function definitions previously marked as TODO in ham-gas.ts were deprecated and will be reimplemented as part of future work. Some locking mechanisms in other parts of the codebase were also revised for simplicity and effectiveness.