Updated HTML Output for XFrameOptionsMode to allow all
The handling of html output has been modified in the gas.js file. Previously, it did not make use of the XFrameOptionsMode. Now, it sets the XFrameOptionsMode to ALLOWALL when the editHtmlOutput option is not being used. This will allow the HTML to be displayed in any iframe regardless of the source.